backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
Pillow>=10.2.0
opencv-python-headless>=4.9.0.80
numpy>=1.26.0
scikit-image>=0.22.0
matplotlib>=3.8.0
openai>=1.14.0
python-dotenv>=1.0.0
